Now rendering the grid meshes as per the LOD walk
Meshes have been scaled to 0-1 regardless of density to make it easier to work with.

Meshes are rendered using instancing, with the instance buffer populated by the LOD walk.
The instance buffer is a fixed size which is based on the number of nodes.

It seems its very efficient, with a max depth of 10, there are about 1.3 million nodes. After the algorithm is done, there are only about 200-300 nodes selected for rendering.

Next step is to introduce a heightmap and get the vertex shader to adjust the vertices.