Added half resolution meshes for rendering tiles outside range sphere.
Previously for quadtree nodes where it was in range of one range level, but only some of its children were in range of the next detail level;
The children which were outside the next detail range were still rendered at full resolution, but with every cell being morphed.
This was inefficient as 3 quarters of the triangles were degenerate.

This commit adds a half resolution mesh for these to use.